+++ src/random_functions.h
@@ -63,7 +63,7 @@ struct rnd_t {
     std::sort(markers.begin(), markers.end());
 
     markers.erase( unique( markers.begin(), markers.end() ), markers.end() );
-    while(markers.size() < number_of_markers) {
+    while(markers.size() < static_cast<size_t>(number_of_markers)) {
       int diff = number_of_markers - markers.size();
       for(int i = 0; i < diff; ++i) {
         markers.push_back(uniform());
